Test and Simulate >> PPO_17. The agent learns adaptive behaviors somehow. Means such reward tuning is effective. Results: https://docs.google.com/presentation/d/1JYPKUz5k5Ncp_WJHWshXA4j_h5Nnnd_D2RlHfh9Taoo/edit?slide=id.g363200e5f87_0_4#slide=id.g363200e5f87_0_4
But some bugs exist, and the training method is not realistic yet, should simulate how much time left. How much text left. The sequentiality is not strong here, need to think about how to do that.

Need to fix the bug: the time condition value was wrongly used as the expected reading time for the specific sentence.